Primer commit con el nuevo repositorio

+++ domain/src/main/java/cat/uvic/teknos/gt3/domain/repositories/BrandRepository.java
@@ -2,5 +2,5 @@ package cat.uvic.teknos.gt3.domain.repositories;
 
 import cat.uvic.teknos.gt3.domain.models.Brand;
 
-public interface BrandRepository extends Repository<Integer, Brand>{
+public interface BrandRepository extends Repository<Long, Brand> {
 }
